Remove findDOMNode injection and inline it
